refactor: split babylon into files
